Lint notes and adjust scope

+++ app/javascript/moduleAdapter/module_adapter.ts
@@ -23,14 +23,14 @@ export default class ModuleAdapter implements IModuleAdapter {
       return false
     }
     try {
-      (this._module.HEAPU8 as Uint8Array).set(encoded, ptr)
+      this._module.HEAPU8.set(encoded, ptr)
     } catch (error: unknown) {
       console.error("range error writing source code to buffer")
       console.error(error?.toString())
     }
     const len = encoded.length
-    if (ptr + len < (this._module.HEAPU8 as Uint8Array).length) {
-      (this._module.HEAPU8 as Uint8Array)[ptr + len] = 0
+    if (ptr + len < this._module.HEAPU8.length) {
+      this._module.HEAPU8[ptr + len] = 0
     }
     const result = this._module._assemble_mmixal(len)
     return result === 0
@@ -78,11 +78,11 @@ export default class ModuleAdapter implements IModuleAdapter {
     if (len === 0) {
       return
     }
-    if (ptr + len >= (this._module.HEAPU8 as Uint8Array).length) {
+    if (ptr + len >= this._module.HEAPU8.length) {
       console.error("overflow: binary executable too large or not terminated")
       return;
     }
-    const bin = (this._module.HEAPU8 as Uint8Array).slice(ptr, ptr + len)
+    const bin = this._module.HEAPU8.slice(ptr, ptr + len)
     if (bin.every((value: number) => value === 0)) {
       console.log("can't simulate an empty executable")
     }
@@ -98,15 +98,15 @@ export default class ModuleAdapter implements IModuleAdapter {
       console.error("pointer may not be null or negative")
       return "simulator error - see logs"
     }
-    if (ptr >= (this._module.HEAPU8 as Uint8Array).length) {
+    if (ptr >= this._module.HEAPU8.length) {
       console.error("pointer out of range")
       return "simulator error - check logs"
     }
     const end = ptr + len
-    if (end >= (this._module.HEAPU8 as Uint8Array).length) {
-      return new TextDecoder().decode((this._module.HEAPU8 as Uint8Array).slice(ptr)) + "--truncated"
+    if (end >= this._module.HEAPU8.length) {
+      return new TextDecoder().decode(this._module.HEAPU8.slice(ptr)) + "--truncated"
     } else {
-      return new TextDecoder().decode((this._module.HEAPU8 as Uint8Array).slice(ptr, end))
+      return new TextDecoder().decode(this._module.HEAPU8.slice(ptr, end))
     }
   }
 }
